refector: 이동 가능 여부를 장기말이 판단하도록 수정

+++ src/main/java/janggi/JanggiGame.java
@@ -35,14 +35,14 @@ public class JanggiGame {
         return new JanggiGame(pieces);
     }
 
-    public void move(final Position start, final Position end) {
-        validateEndPositionPiece(start, end);
-        validatePieceOnPath(start, end);
+    public boolean move(final Position start, final Position end) {
+        if (!pieces.isMoveable(start, end)) {
+            return false;
+        }
         pieces.moveForward(start, end);
-
-        // NOTE: DB에 피스 정보를 저장한다
         janggiDao.deletePieces();
         janggiDao.savePieces(pieces.getPieces());
+        return true;
     }
 
     private void validateEndPositionPiece(Position start, Position end) {
